separate into ec2.go/lightsail.go
