+++ discovery/amazon/ec2.go
@@ -27,7 +27,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws/ec2metadata"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/aws/aws-sdk-go/service/lightsail"
 	"github.com/go-kit/kit/log"
 	"github.com/pkg/errors"
 	"github.com/prometheus/common/config"
@@ -62,30 +61,9 @@ const (
 	ec2LabelSeparator         = ","
 )
 
-const (
-	lightsailLabel                    = model.MetaLabelPrefix + "lightsail_"
-	lightsailLabelAZ                  = lightsailLabel + "availability_zone"
-	lightsailLabelBlueprintID         = lightsailLabel + "blueprint_id"
-	lightsailLabelBundleID            = lightsailLabel + "bundle_id"
-	lightsailLabelInstanceSupportCode = lightsailLabel + "instance_support_code"
-	lightsailLabelInstanceState       = lightsailLabel + "instance_state"
-	lightsailLabelInstanceName        = lightsailLabel + "instance_name"
-	lightsailLabelIPv6Addresses       = lightsailLabel + "ipv6_addresses"
-	lightsailLabelPrivateIP           = lightsailLabel + "private_ip"
-	lightsailLabelPublicIP            = lightsailLabel + "public_ip"
-	lightsailLabelTag                 = lightsailLabel + "tag_"
-	lightsailLabelSeparator           = ","
-)
-
 var (
 	// EC2DefaultSDConfig is the default EC2 SD configuration.
 	EC2DefaultSDConfig = EC2SDConfig{
-		Port:            80,
-		RefreshInterval: model.Duration(60 * time.Second),
-	}
-
-	// LightsailDefaultSDConfig is the default Lightsail SD configuration.
-	LightsailDefaultSDConfig = LightsailSDConfig{
 		Port:            80,
 		RefreshInterval: model.Duration(60 * time.Second),
 	}
@@ -93,7 +71,6 @@ var (
 
 func init() {
 	discovery.RegisterConfig(&EC2SDConfig{})
-	discovery.RegisterConfig(&LightsailSDConfig{})
 }
 
 // Filter is the configuration for filtering EC2 instances.
@@ -115,32 +92,12 @@ type EC2SDConfig struct {
 	Filters         []*EC2Filter   `yaml:"filters"`
 }
 
-// LightsailSDConfig is the configuration for Lightsail based service discovery.
-type LightsailSDConfig struct {
-	Endpoint        string         `yaml:"endpoint"`
-	Region          string         `yaml:"region"`
-	AccessKey       string         `yaml:"access_key,omitempty"`
-	SecretKey       config.Secret  `yaml:"secret_key,omitempty"`
-	Profile         string         `yaml:"profile,omitempty"`
-	RoleARN         string         `yaml:"role_arn,omitempty"`
-	RefreshInterval model.Duration `yaml:"refresh_interval,omitempty"`
-	Port            int            `yaml:"port"`
-}
-
 // Name returns the name of the EC2 Config.
 func (*EC2SDConfig) Name() string { return "ec2" }
-
-// Name returns the name of the Lightsail Config.
-func (*LightsailSDConfig) Name() string { return "lightsail" }
 
 // NewDiscoverer returns a Discoverer for the EC2 Config.
 func (c *EC2SDConfig) NewDiscoverer(opts discovery.DiscovererOptions) (discovery.Discoverer, error) {
 	return NewEC2Discovery(c, opts.Logger), nil
-}
-
-// NewDiscoverer returns a Discoverer for the Lightsail Config.
-func (c *LightsailSDConfig) NewDiscoverer(opts discovery.DiscovererOptions) (discovery.Discoverer, error) {
-	return NewLightsailDiscovery(c, opts.Logger), nil
 }
 
 // UnmarshalYAML implements the yaml.Unmarshaler interface for the EC2 Config.
@@ -171,45 +128,12 @@ func (c *EC2SDConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	return nil
 }
 
-// UnmarshalYAML implements the yaml.Unmarshaler interface for the Lightsail Config.
-func (c *LightsailSDConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	*c = LightsailDefaultSDConfig
-	type plain LightsailSDConfig
-	err := unmarshal((*plain)(c))
-	if err != nil {
-		return err
-	}
-	if c.Region == "" {
-		sess, err := session.NewSession()
-		if err != nil {
-			return err
-		}
-
-		metadata := ec2metadata.New(sess)
-
-		region, err := metadata.Region()
-		if err != nil {
-			return errors.New("Lightsail SD configuration requires a region")
-		}
-		c.Region = region
-	}
-	return nil
-}
-
 // Discovery periodically performs EC2-SD requests. It implements
 // the Discoverer interface.
 type EC2Discovery struct {
 	*refresh.Discovery
 	cfg *EC2SDConfig
 	ec2 *ec2.EC2
-}
-
-// Discovery periodically performs Lightsail-SD requests. It implements
-// the Discoverer interface.
-type LightsailDiscovery struct {
-	*refresh.Discovery
-	cfg       *LightsailSDConfig
-	lightsail *lightsail.Lightsail
 }
 
 // NewEC2Discovery returns a new EC2Discovery which periodically refreshes its targets.
@@ -223,23 +147,6 @@ func NewEC2Discovery(conf *EC2SDConfig, logger log.Logger) *EC2Discovery {
 	d.Discovery = refresh.NewDiscovery(
 		logger,
 		"ec2",
-		time.Duration(d.cfg.RefreshInterval),
-		d.refresh,
-	)
-	return d
-}
-
-// NewLightsailDiscovery returns a new LightsailDiscovery which periodically refreshes its targets.
-func NewLightsailDiscovery(conf *LightsailSDConfig, logger log.Logger) *LightsailDiscovery {
-	if logger == nil {
-		logger = log.NewNopLogger()
-	}
-	d := &LightsailDiscovery{
-		cfg: conf,
-	}
-	d.Discovery = refresh.NewDiscovery(
-		logger,
-		"lightsail",
 		time.Duration(d.cfg.RefreshInterval),
 		d.refresh,
 	)
@@ -276,33 +183,6 @@ func (d *EC2Discovery) ec2Client() (*ec2.EC2, error) {
 	}
 
 	return d.ec2, nil
-}
-
-func (d *LightsailDiscovery) lightsailClient() (*lightsail.Lightsail, error) {
-	if d.lightsail != nil {
-		return d.lightsail, nil
-	}
-
-	creds := credentials.NewStaticCredentials(d.cfg.AccessKey, string(d.cfg.SecretKey), "")
-	if d.cfg.AccessKey == "" && d.cfg.SecretKey == "" {
-		creds = nil
-	}
-
-	sess, err := session.NewSessionWithOptions(session.Options{
-		Config: aws.Config{
-			Endpoint:    &d.cfg.Endpoint,
-			Region:      &d.cfg.Region,
-			Credentials: creds,
-		},
-		Profile: d.cfg.Profile,
-	})
-	if err != nil {
-		return nil, errors.Wrap(err, "could not create aws session")
-	}
-
-	d.lightsail = lightsail.New(sess)
-
-	return d.lightsail, nil
 }
 
 func (d *EC2Discovery) refresh(ctx context.Context) ([]*targetgroup.Group, error) {
@@ -417,73 +297,6 @@ func (d *EC2Discovery) refresh(ctx context.Context) ([]*targetgroup.Group, error
 			d.ec2 = nil
 		}
 		return nil, errors.Wrap(err, "could not describe instances")
-	}
-	return []*targetgroup.Group{tg}, nil
-}
-
-func (d *LightsailDiscovery) refresh(ctx context.Context) ([]*targetgroup.Group, error) {
-	lightsailClient, err := d.lightsailClient()
-	if err != nil {
-		return nil, err
-	}
-
-	tg := &targetgroup.Group{
-		Source: d.cfg.Region,
-	}
-
-	input := &lightsail.GetInstancesInput{}
-
-	output, err := lightsailClient.GetInstancesWithContext(ctx, input)
-	if err != nil {
-		if awsErr, ok := err.(awserr.Error); ok && (awsErr.Code() == "AuthFailure" || awsErr.Code() == "UnauthorizedOperation") {
-			d.lightsail = nil
-		}
-		return nil, errors.Wrap(err, "could not describe instances")
-	}
-
-	for _, inst := range output.Instances {
-		if inst.PrivateIpAddress == nil {
-			continue
-		}
-
-		labels := model.LabelSet{
-			lightsailLabelInstanceSupportCode: model.LabelValue(*inst.SupportCode),
-		}
-
-		labels[lightsailLabelPrivateIP] = model.LabelValue(*inst.PrivateIpAddress)
-		addr := net.JoinHostPort(*inst.PrivateIpAddress, fmt.Sprintf("%d", d.cfg.Port))
-		labels[model.AddressLabel] = model.LabelValue(addr)
-
-		if inst.PublicIpAddress != nil {
-			labels[lightsailLabelPublicIP] = model.LabelValue(*inst.PublicIpAddress)
-		}
-
-		if len(inst.Ipv6Addresses) > 0 {
-			var ipv6addrs []string
-			for _, ipv6addr := range inst.Ipv6Addresses {
-				ipv6addrs = append(ipv6addrs, *ipv6addr)
-			}
-			labels[lightsailLabelIPv6Addresses] = model.LabelValue(
-				lightsailLabelSeparator +
-					strings.Join(ipv6addrs, lightsailLabelSeparator) +
-					lightsailLabelSeparator)
-		}
-
-		labels[lightsailLabelAZ] = model.LabelValue(*inst.Location.AvailabilityZone)
-		labels[lightsailLabelBlueprintID] = model.LabelValue(*inst.BlueprintId)
-		labels[lightsailLabelBundleID] = model.LabelValue(*inst.BundleId)
-		labels[lightsailLabelInstanceName] = model.LabelValue(*inst.Name)
-		labels[lightsailLabelInstanceState] = model.LabelValue(*inst.State.Name)
-
-		for _, t := range inst.Tags {
-			if t == nil || t.Key == nil || t.Value == nil {
-				continue
-			}
-			name := strutil.SanitizeLabelName(*t.Key)
-			labels[lightsailLabelTag+model.LabelName(name)] = model.LabelValue(*t.Value)
-		}
-
-		tg.Targets = append(tg.Targets, labels)
 	}
 	return []*targetgroup.Group{tg}, nil
 }
